1.3 -> 1.4 changes (+patch)

+++ src/main.c
@@ -113,6 +113,7 @@ struct nl_sock *setup_netlink_socket()
 	struct nl_sock *sd;
 	int family;
 
+
 	sd = nl_socket_alloc();
 
 	genl_connect(sd);
@@ -332,7 +333,7 @@ void handle_dm_alert_msg(struct netlink_message *msg, int err)
 			printf ("%d drops at location %p\n", alert->points[i].count, location);
 		else
 			printf ("%d drops at %s+%llx (%p)\n",
-				alert->points[i].count, res.symbol, res.offset, location);
+				alert->points[i].count, res.symbol, (long long unsigned)res.offset, location);
 		acount++;
 		if (alimit && (acount == alimit)) {
 			printf("Alert limit reached, deactivating!\n");
